Corrigida a representação da matriz (debug). Acrescentada uma função para imprimir o vetor (debug). Corrigida o grafo fixo (debug) Corrigida a implementação da inicialização de v_t e d.

+++ src/prim/prim_st.h
@@ -19,4 +19,4 @@
 #include "../graph/graph.h"
 #endif
 
-int *prim_mst(int array_size, int graph_size, int graph_root);
+float *prim_mst(int array_size, int graph_size, int graph_root);
